+++ crawjud/crawjud_manager/runner.py
@@ -8,9 +8,9 @@ from clear import clear
 from termcolor import colored
 from tqdm import tqdm
 
+from crawjud._types import app_name
 from crawjud.config import StoreThread, running_servers
 from crawjud.core.watch import monitor_log
-from crawjud.types import app_name
 from crawjud.utils.gen_seed import worker_name_generator
