+++ crawjud/decorators/checks.py
@@ -7,8 +7,8 @@ from flask_sqlalchemy import SQLAlchemy
 from quart import Response, flash, make_response, redirect, session, url_for
 from quart import current_app as app
 
+from crawjud._types import AnyStr, WrappedFnReturnT
 from crawjud.models import Users
-from crawjud.types import AnyStr, WrappedFnReturnT
 
 
 def check_privilegies(func: Callable[[], Response]) -> WrappedFnReturnT:
