+++ crawjud/forms/credentials.py
@@ -7,7 +7,7 @@ from quart_wtf import QuartForm
 from wtforms import PasswordField, SelectField, StringField, SubmitField
 from wtforms.validators import DataRequired
 
-from crawjud.types import AnyType, T
+from crawjud._types import AnyType, T
 
 file_allowed = FileAllowed(["pfx", 'Apenas arquivos ".pfx" são permitidos!'])
